[SID:3619] fix: CI shard 4 60초 가드 재발(test_3498) — weights 갱신+파일 분리
2026-09-07 06:29Z, #3971 run 34089668525 shard 4: test_3498_generation_
budget.py(당시 25테스트·등재 42s) 격리실행이 100s > 88.0s 가드를 넘겨 Backend
pytest 잡까지 실패 → PO rerun. 즉시 rerun(같은 run)은 40.30s로 통과 — 재현이
러너 경합에 크게 좌우됨을 확認(#3604 test_3502 실사례와 같은 클래스).

## 원인 판별(로컬 --durations 실측)
23개 실DB 테스트가 각 0.32~0.47s로 균등 분포 — #3502처럼 "app.main import
유무로 갈리는 무거운 조각 하나"가 없다(첫 테스트만 app.main 최초 import
비용을 안고 2.2~4.2s, 그 외는 고르다). 대신 파일이 이미 갖고 있던 8절 구분
(AC1~AC5·조각①·조각⑤·PUT 3종) 경계를 그대로 써서 절반으로 갈랐다 — 테스트를
절 중간에서 자르지 않는다.

## 분리
- 원 파일: AC1~AC5(12테스트) 유지.
- 신설 `test_3498_generation_budget_evidence_and_config.py`: 조각①(GET
  /generation-budget)·조각⑤(evidence payload 검증)·PUT 3종(11테스트). 세팅
  헬퍼는 test_3471_org_content_rules_lint.py 재사용(#3502 분리 선례와 동형
  관례) + 이 절만 쓰는 로컬 헬퍼(_seed_evidence_agent 등)는 원 파일에서 제거
  (분리 후 dead code였다)하고 이 파일에 그대로 둠.
- 테스트 수·assertion 전부 보존(원 23개 = 12+11), 두 파일 합쳐 로컬 12.42s.

## weights 갱신
run 34089668525의 100s(25테스트)를 12/25·11/25 비율로 안분(52.0s·48.0s) —
40.30s(rerun) 기준 안분(21.0s/19.3s)이 아니라 100s(경합) 기준을 택해 재발
여유를 더 크게 뒀다. source에 run id·양쪽 elapsed·안분 근거 명기(3604 관례).

## 검증
- 두 파일 합쳐 23 passed(로컬, 12.42s).
- lint_destructive_schema_weights_registered.py: 294 discovered · 294
  registered · 신규 미등재 0건.
